requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
greenlet==1.1.2
idna==3.3
jedi==0.18.1
msgpack==1.0.3
neovim==0.3.1
parso==0.8.3
pycparser==2.21
PyMySQL==1.0.2
pynvim==0.4.3
python-dotenv==0.20.0
requests==2.27.1
tqdm==4.64.0
urllib3==1.26.9
pydantic==1.9.1
